+++ backends/simulation/simulation/ResultStruct.h
@@ -6,9 +6,9 @@ struct SimulationResultStruct
 	SimulationResultStruct() : 
 		reliability(1.0), meanAvailability(1.0), mttf(-1.0), nRounds(0), nFailures(0), duration(0.0) {};
 
-	long double reliability;
-	long double meanAvailability;
-	long double mttf;
+	double reliability;
+	double meanAvailability;
+	double mttf;
 	unsigned int nRounds;
 	unsigned int nFailures;
 	double duration;
